(GP-2.2): Created a Hash Function
Implements SHA-1 hashing for string inputs. The SHA-1 is created in the hashFunction method. This provides a way to generate unique identifiers for strings. It uses a try catch in the process of creating the hash function. The hashing functionality is tested in the main method for demonstration. The implementation is based on the SHA-1 initialization format from GeeksforGeeks. No helper methods were used.