Adds index update functionality
THIS CODE IS WRONG SO FAR BUT WILL FIX LATER
Implements the ability to update the index file with the hash and file path of a given file.

This allows the system to track changes to files and their corresponding hashes, enabling future commit and rollback functionalities.

Adds a test case to verify the index update functionality.